Move `GenerateNotebook` to notebook_renderer.py
